Major: 0.3.36
New Features:
1. Use a brand new way of set system and user data.

+++ src/core/knpluginmanager.cpp
@@ -88,7 +88,7 @@ KNPluginManager::KNPluginManager(QObject *parent) :
 void KNPluginManager::backupWindowGeometry()
 {
     //Backup the window state.
-    m_global->setSystemData("windowState", m_mainWindow->windowState());
+    m_global->setSystemData("windowState", (int)m_mainWindow->windowState());
     //Window position.
     m_global->setSystemData("windowX", m_mainWindow->geometry().x());
     m_global->setSystemData("windowY", m_mainWindow->geometry().y());
